♻️ refactor: refactor the plugin and tool slice
